HTTP basic auth user name encoding in UTF-8
This should lead to more compatibility with the
chemistry cmis server implementation and now should
act equal to the java client lib implementation.

This change makes it possible to log in with a user name
containing umlauts etc.

The past usage of ISO-8859-1 encoding was not compatible
with the chemistry server and the rfc2617 does not
describe any encoding to be used.